% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicycle_theft.R
\name{bicycle_theft}
\alias{bicycle_theft}
\title{Access the statistics about bicycle theft of the State of Rio de Janeiro}
\usage{
bicycle_theft()
}
\value{
a dataframe
}
\description{
Returns monthly data on police occurrences involving bicycle theft in the State of Rio de Janeiro in the form of a dataframe.
}
\examples{
bicycle_theft()

}
