% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_fleet.R
\name{car_fleet}
\alias{car_fleet}
\title{Access data from car fleet}
\usage{
car_fleet(by)
}
\arguments{
\item{by}{"month" or "year". character.}
}
\value{
a dataframe.
}
\description{
Returns data about car fleet in State of Rio de Janeiro in the form of a dataframe.
}
\examples{
car_fleet(by = "month")

}
