% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gun_seizure.R
\name{gun_seizure}
\alias{gun_seizure}
\title{Access data from gun seizure in State of Rio de Janeiro}
\usage{
gun_seizure(gun_type)
}
\arguments{
\item{gun_type}{selects the gun type: "firearms" or "edged_weapons". character.}
}
\value{
a dataframe
}
\description{
Returns monthly data about gun seizure in State of Rio de Janeiro in the form of a dataframe.
}
\examples{
\donttest{gun_seizure(gun_type = "firearms")}

}
