% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gun_seizure_historical.R
\name{gun_seizure_historical}
\alias{gun_seizure_historical}
\title{Access historical data from seizure of firearms in State of Rio de Janeiro}
\usage{
gun_seizure_historical(by)
}
\arguments{
\item{by}{selects the gun type: "firearms" or "edged_weapons". character.}
}
\value{
a dataframe
}
\description{
Returns historical data about seizure of firearms in State of Rio de Janeiro in the form of a dataframe.
}
\details{
Data availability:
total for the state from 2000 to 2006.
total for police station area from 2003 to 2006.
}
\examples{
gun_seizure_historical(by = "police_station_area")

}
