% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_stats.R
\name{monthly_stats}
\alias{monthly_stats}
\title{Access the public security statistics of the State of Rio de Janeiro by month}
\usage{
monthly_stats(by, value = "standard")
}
\arguments{
\item{by}{selects the spatial division of the data. It might be: "police_station_area", "municipality" or "state". character.}

\item{value}{allows you to choose whether the values will be in absolute numbers ("standard") or per 100,000 inhabitants ("per_100k"). character.}
}
\value{
a dataframe
}
\description{
Returns monthly data on police occurrences in the State of Rio de Janeiro in the form of a dataframe.
}
\details{
To see the dictionary of variables, use the function monthly_stats_dictionary().
}
\examples{
\donttest{monthly_stats(by = "police_station_area")}

}
