% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_stats_dictionary.R
\name{monthly_stats_dictionary}
\alias{monthly_stats_dictionary}
\title{Access the dictionaries of public security statistics of the State of Rio de Janeiro by month}
\usage{
monthly_stats_dictionary(by, value = "standard")
}
\arguments{
\item{by}{selects the spatial division of the dictionary. It might be: "cisp", "municipality" or "state". character.}

\item{value}{selects the type of value: absolute numbers ("standard") or per 100,000 inhabitants ("per_100k"). character.}
}
\value{
a dataframe
}
\description{
Returns the data dictionary of the police occurrences in the State of Rio de Janeiro in the form of a dataframe.
}
\examples{
monthly_stats_dictionary(by = "state", value = "per_100k")

}
