% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Access population data}
\usage{
population(data)
}
\arguments{
\item{data}{selects the data: "cisp_monthly", "cisp_yearly", "muni_monthly", "muni_yearly", "state_monthly", "state_yearly", "upp_2010" or "upp_projection". character.}
}
\value{
a dataframe.
}
\description{
Returns data about population by state, municipality, CISP (the same as police station area) or Pacifying Police Unit.
}
\examples{
population(data = "cisp_monthly")

}
