% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_aisp.R
\docType{data}
\name{spatial_aisp}
\alias{spatial_aisp}
\title{Spatial Data from Integrated Public Safety Area (AISP)}
\format{
\subsection{\code{spatial_aisp}}{

A spatial data frame
\describe{
\item{aisp}{Integrated Public Safety Areas (AISP) code}
\item{risp}{Integrated Public Safety Regions (RISP) code}
\item{geometry}{Spatial vector data of AISP}
}
}
}
\source{
\url{https://www.ispdados.rj.gov.br/Arquivos/AISPshp.rar}
}
\usage{
spatial_aisp
}
\description{
This is the spatial data from the Integrated Public Safety Area (AISP) of the State of Rio de Janeiro.
}
\keyword{datasets}
