% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cisp.R
\docType{data}
\name{spatial_cisp}
\alias{spatial_cisp}
\title{Spatial Data from Integrated Public Safety Circumscriptions (CISP)}
\format{
\subsection{\code{spatial_cisp}}{

A spatial data frame
\describe{
\item{dp}{id number}
\item{aisp}{Integrated Public Safety Areas (AISP) code}
\item{geometry}{Spatial vector data of CISP}
}
}
}
\source{
\url{https://www.ispdados.rj.gov.br/Arquivos/CISPshp.rar}
}
\usage{
spatial_cisp
}
\description{
This is the spatial data from the Integrated Public Safety Circumscriptions (CISP) of the State of Rio de Janeiro.
}
\keyword{datasets}
