% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_info.R
\name{spatial_info}
\alias{spatial_info}
\title{Access the information about spatial data}
\usage{
spatial_info(data)
}
\arguments{
\item{data}{it may be "code_equivalence" or "area_km2"}
}
\value{
a dataframe
}
\description{
Returns data about the spatial data in the package. If the data argument is "code_equivalence",
it will show the equivalence between AISP, CISP and RISP codes, neighborhoods and cities. If the
data argument is "area_km2", it returns a dataframe with the areas of AISP, CISP and RISP by square kilometers.
}
\examples{
\dontrun{spatial_info(data = "code_equivalence")}

}
