% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_regions_rj.R
\docType{data}
\name{spatial_regions_rj}
\alias{spatial_regions_rj}
\title{Spatial Data from Regions of the State of Rio de Janeiro}
\format{
\subsection{\code{spatial_regions_rj}}{

A spatial data frame
\describe{
\item{id1}{id number}
\item{regiao}{region name}
\item{geometry}{Spatial vector data of the regions}
}
}
}
\source{
\url{https://www.ispdados.rj.gov.br/Arquivos/RegioesSHP.rar}
}
\usage{
spatial_regions_rj
}
\description{
This is the spatial data from the State of Rio de Janeiro.
}
\keyword{datasets}
