% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_risp.R
\docType{data}
\name{spatial_risp}
\alias{spatial_risp}
\title{Spatial Data from Integrated Public Safety Regions (RISP)}
\format{
\subsection{\code{spatial_risp}}{

A spatial data frame
\describe{
\item{risp}{risp number}
\item{geometry}{Spatial vector data of RISP}
}
}
}
\source{
\url{https://www.ispdados.rj.gov.br/Arquivos/RISPshp.rar}
}
\usage{
spatial_risp
}
\description{
This is the spatial data from the Integrated Public Safety Regions (RISP) of the State of Rio de Janeiro.
}
\keyword{datasets}
