% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_upp.R
\docType{data}
\name{spatial_upp}
\alias{spatial_upp}
\title{Spatial data of Pacifying Police Units (UPP) - 2017 Limits}
\format{
\subsection{\code{spatial_upp}}{

A spatial data frame
\describe{
\item{nome}{UPP name}
\item{nomeabrev}{UPP abbreviate name}
\item{id_upp}{UPP id number}
\item{datacriaca}{Creation date}
\item{resolucaoc}{resolution number}
\item{geometry}{Spatial vector data of UPPs limits in 2017}
}
}
}
\source{
\url{https://www.ispdados.rj.gov.br/Arquivos/UPPshp.rar}
}
\usage{
spatial_upp
}
\description{
This is the spatial data from the territorial limits of the Pacifying Police Units (UPPs) in the year of 2017.
}
\keyword{datasets}
