% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upp_data.R
\name{upp_data}
\alias{upp_data}
\title{Access data from Pacifying Police Units (UPPs)}
\usage{
upp_data(data)
}
\arguments{
\item{data}{selects the data: "stats", "dates" or "area_m2". character.}
}
\value{
a dataframe.
}
\description{
Returns data about Pacifying Police Units in the form of a dataframe.
}
\examples{
upp_data(data = "area_m2")

}
