% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_stats.R
\name{yearly_stats}
\alias{yearly_stats}
\title{Access the public security statistics of the State of Rio de Janeiro by year}
\usage{
yearly_stats(by)
}
\arguments{
\item{by}{selects the spatial division of the data. It might be: "municipality" or "state". character.}
}
\value{
a dataframe
}
\description{
Returns yearly data on police occurrences in the State of Rio de Janeiro in the form of a dataframe.
}
\details{
Values per 100,000 inhabitants.

To see the dictionary of variables, use the function yearly_stats_dictionary().
}
\examples{
yearly_stats(by = "state")

}
