% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_stats_dictionary.R
\name{yearly_stats_dictionary}
\alias{yearly_stats_dictionary}
\title{Access the dictionaries of public security statistics of the State of Rio de Janeiro by year}
\usage{
yearly_stats_dictionary(by)
}
\arguments{
\item{by}{selects the spatial division of the dictionary. It might be: "municipality" or "state". character.}
}
\value{
a dataframe
}
\description{
Returns the data dictionary of the police occurrences in the State of Rio de Janeiro in the form of a dataframe.
}
\examples{
yearly_stats_dictionary(by = "state")

}
