% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimes_against_life.R
\name{crimes_against_life}
\alias{crimes_against_life}
\title{Access the public security statistics about crimes against life in the State of Rio de Janeiro}
\usage{
crimes_against_life(type)
}
\arguments{
\item{type}{allows you to select the type of violent crime: "femicide", "violent_lethality", "violent_lethality_elucidation_rate" or "officers_killed_on_duty". character.}
}
\value{
a dataframe
}
\description{
Returns monthly data about crimes against life in the State of Rio de Janeiro in the form of a dataframe.
}
\details{
The data available are on femicide, violent lethality, the rate of elucidation of cases of violent lethality or death of police officers on duty.
}
\examples{
\dontrun{crimes_against_life(type = "femicide")}

}
