% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gun_seizure_historical.R
\name{gun_seizure_historical}
\alias{gun_seizure_historical}
\title{Access historical data from seizure of firearms in State of Rio de Janeiro}
\usage{
gun_seizure_historical(by)
}
\arguments{
\item{by}{selects the spatial division of the data: "police_station_area" or "state". character.}
}
\value{
a dataframe
}
\description{
Returns monthly historical data about seizure of firearms in State of Rio de Janeiro in the form of a dataframe.
}
\details{
Data availability from 2000 to 2006 for the State and from 2003 to 2006 for the Police Station Area.
}
\examples{
gun_seizure_historical(by = "police_station_area")

}
